Product/requirements.txt
clarifai_grpc==8.0.0
firebase_admin==5.2.0
Flask==2.0.2
Flask_Cors==3.0.10
Flask_SQLAlchemy==2.5.1
gevent==21.12.0
mailjet_rest==1.3.4
Pyrebase4==4.5.0
tornado==6.1
Twisted==22.2.0
requests==2.27.1
mysql-connector-python==8.0.27
